refactor bottles list in to component
Going to use this later on in the admin
